refactor: updating LoginPage and RegisterPage
